requirements.txt
asyncio
requests
sql
aiogram
bs4
beautifulsoup4
python-dotenv